requirements.txt
pyTelegramBotAPI
google-generativeai
python-dotenv
Pillow
PyPDF2
speedtest-cli
